Add support for not ignoring stdout when performing git actions.
This will be needed when adding support for git add/reset --patch (it requires
stdout to print the hunks).

Also, rename the ignore_errors parameter to ignore_stderr to make it match the
newly added parameter ignore_stdout. The new name is more precise as we are not
ignoring errors, only stderr.